Add aula_08

+++ aula_08/blogpessoal/src/main/java/br/org/generation/blogpessoal/controller/TemaController.java
@@ -1,6 +1,7 @@
 package br.org.generation.blogpessoal.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -14,50 +15,70 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import br.org.generation.blogpessoal.model.Tema;
 import br.org.generation.blogpessoal.repository.TemaRepository;
+
 
 @RestController
 @RequestMapping("/temas")
 @CrossOrigin(origins = "*", allowedHeaders = "*")
 public class TemaController {
-
-	@Autowired
+    
+    @Autowired
 	private TemaRepository temaRepository;
 
 	@GetMapping
 	public ResponseEntity<List<Tema>> getAll() {
 		return ResponseEntity.ok(temaRepository.findAll());
-
+	
 	}
 
 	@GetMapping("/{id}")
 	public ResponseEntity<Tema> getById(@PathVariable long id) {
 		return temaRepository.findById(id)
-			.map(resp -> ResponseEntity.ok(resp))
-			.orElse(ResponseEntity.notFound().build());
+				.map(resp -> ResponseEntity.ok(resp))
+				.orElse(ResponseEntity.notFound().build());
 	}
-
+	
 	@GetMapping("/descricao/{descricao}")
 	public ResponseEntity<List<Tema>> getByDescricao(@PathVariable String descricao) {
 		return ResponseEntity.ok(temaRepository.findAllByDescricaoContainingIgnoreCase(descricao));
 	}
-
+	
 	@PostMapping
-	public ResponseEntity<Tema> postTema(@RequestBody Tema tema) {
-		return ResponseEntity.status(HttpStatus.CREATED).body(temaRepository.save(tema));
+	public ResponseEntity<Tema> postTema(@RequestBody Tema tema){
+		return ResponseEntity.status(HttpStatus.CREATED)
+				.body(temaRepository.save(tema));
 	}
 
 	@PutMapping
-	public ResponseEntity<Tema> putTema(@RequestBody Tema tema) {
-		return ResponseEntity.status(HttpStatus.OK).body(temaRepository.save(tema));
-
+	public ResponseEntity<Tema> putTema(@RequestBody Tema tema){
+		
+		Optional<Tema> temaUpdate = temaRepository.findById(tema.getId());
+		
+		if (temaUpdate.isPresent()) {
+			return ResponseEntity.status(HttpStatus.OK).body(temaRepository.save(tema));
+		}else{
+			throw new ResponseStatusException(
+		          	HttpStatus.NOT_FOUND, "Tema não encontrado!", null);
+		}
+			
 	}
-
+	
 	@DeleteMapping("/{id}")
 	public void deleteTema(@PathVariable long id) {
-		temaRepository.deleteById(id);
+
+		Optional<Tema> tema = temaRepository.findById(id);
+		
+		if (tema.isPresent()) {
+			temaRepository.deleteById(id);
+		}else{
+			throw new ResponseStatusException(
+		          	HttpStatus.NOT_FOUND, "Tema não encontrado!", null);
+		}
 	}
 
 }
+
